removed unnecessary db call and null check before settings delete in zettle

+++ src/Plugins/Nop.Plugin.Misc.Zettle/Data/InventoryBalanceMigration.cs
@@ -40,7 +40,8 @@ public class InventoryBalanceMigration : MigrationBase
 
         //delete settings
         var setting = await _settingService.GetSettingAsync($"{nameof(ZettleSettings)}.InventoryTrackingIds");
-        await _settingService.DeleteSettingAsync(setting);
+        if(setting is not null)
+            await _settingService.DeleteSettingAsync(setting);
     }
 
     /// <summary>
